justify the coming data
